Use Sanitizer checks and fix error + fix some more critical  clang-tidy errors

+++ src/grid.cpp
@@ -1,15 +1,15 @@
 #include "grid.hpp"
 #include "renderer.hpp"
 
-Grid::Grid(Point offset, int tile_size) : m_offset{ offset - Point{ 0, invisible_rows * tile_size }}, m_tile_size{ tile_size } {
+Grid::Grid(Point offset, usize tile_size) : m_offset{ offset - Point{ 0, static_cast<int>(invisible_rows * tile_size) }}, m_tile_size{ tile_size } {
 }
 
 [[nodiscard]] Point Grid::tile_size() const {
-    return Point{ m_tile_size, m_tile_size };
+    return Point{ static_cast<int>(m_tile_size), static_cast<int>(m_tile_size) };
 }
 
 [[nodiscard]] Point Grid::to_screen_coords(Point grid_coords) const {
-    return m_offset + grid_coords * m_tile_size;
+    return m_offset + grid_coords * static_cast<int>(m_tile_size);
 }
 
 void Grid::render(const ServiceProvider& service_provider) const {
@@ -49,19 +49,20 @@ void Grid::draw_playing_field_background(const ServiceProvider& service_provider
 }
 
 void Grid::draw_background(const ServiceProvider& service_provider, Rect grid_rect) const {
-    const auto top_left = m_offset + Point{ 0, m_tile_size * invisible_rows } + grid_rect.top_left * m_tile_size;
-    const auto bottom_right = top_left + Point{ grid_rect.width(), grid_rect.height() } * m_tile_size;
+    const auto top_left = m_offset + Point{ 0, static_cast<int>(m_tile_size * invisible_rows) }
+                          + grid_rect.top_left * static_cast<int>(m_tile_size);
+    const auto bottom_right = top_left + Point{ grid_rect.width(), grid_rect.height() } * static_cast<int>(m_tile_size);
     service_provider.renderer().draw_rect_filled(Rect{ top_left, bottom_right }, background_color);
 
-    for (int column = 0; column <= grid_rect.width(); ++column) {
-        const auto start = top_left + Point{ column * m_tile_size, 0 };
-        const auto end = Point{ start.x, start.y + grid_rect.height() * m_tile_size };
+    for (usize column = 0; column <= static_cast<usize>(grid_rect.width()); ++column) {
+        const auto start = top_left + Point{ static_cast<int>(column * m_tile_size), 0 };
+        const auto end = Point{ start.x, static_cast<int>(start.y + grid_rect.height() * m_tile_size) };
         service_provider.renderer().draw_line(start, end, grid_color);
         service_provider.renderer().draw_line(start - Point{ 1, 0 }, end - Point{ 1, 0 }, grid_color);
     }
 
-    for (int row = 0; row <= grid_rect.height(); ++row) {
-        const auto start = top_left + Point{ 0, row * m_tile_size };
+    for (usize row = 0; row <= static_cast<usize>(grid_rect.height()); ++row) {
+        const auto start = top_left + Point{ 0, static_cast<int>(row * m_tile_size) };
         const auto end = Point{ bottom_right.x, start.y };
         service_provider.renderer().draw_line(start, end, grid_color);
         service_provider.renderer().draw_line(start - Point{ 0, 1 }, end - Point{ 0, 1 }, grid_color);
@@ -69,8 +70,8 @@ void Grid::draw_background(const ServiceProvider& service_provider, Rect grid_re
 
     const auto outline_top_left = top_left - Point{ 2, 2 };
     const auto outline_bottom_right = Point{
-        top_left.x + grid_rect.width() * m_tile_size + 1,
-        top_left.y + grid_rect.height() * m_tile_size + 1,
+        static_cast<int>(top_left.x + grid_rect.width() * m_tile_size + 1),
+        static_cast<int>(top_left.y + grid_rect.height() * m_tile_size + 1),
     };
 
     const Rect outline_rect = Rect{
